One Hot Debugging

+++ model/src/dataset.py
@@ -15,8 +15,8 @@ class DataSet(object):
             unpickler = pickle._Unpickler(file)
             unpickler.encoding = 'latin1'
             contents = unpickler.load()
-            X, Y = np.asarray(contents['data'], dtype=np.float32), np.asarray(contents['labels'], dtype=np.float32)
-            one_hot = np.zeros(Y.size, Y.max() + 1)
+            X, Y = np.asarray(contents['data'], dtype=np.float32), np.asarray(contents['labels'])
+            one_hot = np.zeros((Y.size, Y.max() + 1))
             one_hot[np.arange(Y.size), Y] = 1
             return X, one_hot
